pyproject.toml
[tool.poetry]
name = "patent-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12" 
pandas = "2.1"
dask = "^2023.1.0"
ipykernel = "^6.20.1"
en_core_web_md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1.tar.gz#egg=en_core_web_sm"}
en_core_web_lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1.tar.gz#egg=en_core_web_lg"}
# en_core_web_trf = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1.tar.gz#egg=en_core_web_trf"}
notebook = "^6.5.2"
pyarrow = "^10.0.1"
openpyxl = "^3.0.10"
pyalex = "^0.15"
matplotlib = "^3.6.3"
seaborn = "^0.12.2"
pip = "^22.3.1"
torch = "^1.13.1"
spacy-transformers = "^1.2.1"
jupyter = "^1.0.0"
requests-cache = "^0.9.8"
jupyterlab = "^3.6.1"
xlsxwriter = "^3.0.8"
scipy = "^1.10.0"
xlrd = "^2.0.1"
python-dotenv = "^0.21.1"
click = "^8.1.3"
bokeh = "^3.0.3"
scikit-learn = "^1.2.1"
statsmodels = "^0.13.5"
pylint = "^2.17.0"
flake8 = "^6.0.0"
ruff = "^0.9.1"
polars = "^0.16.12"
sentence-transformers = "^2.2.2"
linearmodels = "^5.4"
patsy = "^0.5.6"
duckdb = "^1.0.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[tool.ruff]
fixable = ["A", "B", "C", "D", "E", "F"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


